django_backend/requirements.txt
Django~=4.0  
djangorestframework~=3.13 
django-cors-headers~=3.10  
Pillow~=9.0  
django-storages google-cloud-storage